+++ include/dbldbl.h
@@ -43,6 +43,8 @@
  *     observed maximum relative error.
  */
 
+#include <quda_mathfuncs.h>
+
 #if !defined(DBLDBL_H_)
 #define DBLDBL_H_
 
@@ -226,7 +228,7 @@ __device__ __forceinline__ dbldbl sqrt_dbldbl (dbldbl a)
 {
     dbldbl t, z;
     double e, y, s, r;
-    r = rsqrt (a.y);
+    r = quda::rsqrt (a.y);
     if (a.y == 0.0) r = 0.0;
     y = __dmul_rn (a.y, r);
     s = __fma_rn (y, -y, a.y);
@@ -253,7 +255,7 @@ __device__ __forceinline__ dbldbl rsqrt_dbldbl (dbldbl a)
 {
     dbldbl z;
     double r, s, e;
-    r = rsqrt (a.y);
+    r = quda::rsqrt (a.y);
     e = __dmul_rn (a.y, r);
     s = __fma_rn (e, -r, 1.0);
     e = __fma_rn (a.y, r, -e);
