+++ include/gauge_field_order.h
@@ -10,6 +10,7 @@
 #include <type_traits>
 
 #include <register_traits.h>
+#include <quda_mathfuncs.h>
 #include <convert.h>
 #include <complex_quda.h>
 #include <quda_matrix.h>
@@ -1534,7 +1535,7 @@ namespace quda {
           real row_sum_inv = static_cast<real>(1.0) / row_sum;
 
           real diff = u0_inv * u0_inv - row_sum;
-          real U00_mag = diff > 0.0 ? diff * rsqrt(diff) : static_cast<real>(0.0);
+          real U00_mag = diff > 0.0 ? diff * quda::rsqrt(diff) : static_cast<real>(0.0);
 
           out[0] *= U00_mag;
 
@@ -1545,7 +1546,7 @@ namespace quda {
           column_sum += out[3].imag() * out[3].imag();
 
           diff = u0_inv * u0_inv - column_sum;
-          real U20_mag = diff > 0.0 ? diff * rsqrt(diff) : static_cast<real>(0.0);
+          real U20_mag = diff > 0.0 ? diff * quda::rsqrt(diff) : static_cast<real>(0.0);
 
           out[6] *= U20_mag;
