+++ include/kernels/gauge_random.cuh
@@ -1,4 +1,5 @@
 #include <quda_matrix.h>
+#include <quda_mathfuncs.h>
 #include <gauge_field_order.h>
 #include <index_helper.cuh>
 #include <random_helper.h>
@@ -52,15 +53,16 @@ namespace quda {
     for (int i = 0; i < 4; ++i) {
       phi[i] = 2.0 * M_PI * rand1[i];
       radius[i] = sqrt(-log(rand2[i]));
-      sincos(phi[i], &temp2[i], &temp1[i]);
+      quda::sincos(phi[i], &temp2[i], &temp1[i]);
       temp1[i] *= radius[i];
       temp2[i] *= radius[i];
     }
 
     // construct Anti-Hermitian matrix
-    ret(0, 0) = complex<real>(0.0, temp1[2] + rsqrt(3.0) * temp2[3]);
-    ret(1, 1) = complex<real>(0.0, -temp1[2] + rsqrt(3.0) * temp2[3]);
-    ret(2, 2) = complex<real>(0.0, -2.0 * rsqrt(3.0) * temp2[3]);
+    const real rsqrt_3 = quda::rsqrt(3.0);
+    ret(0, 0) = complex<real>(0.0, temp1[2] + rsqrt_3 * temp2[3]);
+    ret(1, 1) = complex<real>(0.0, -temp1[2] + rsqrt_3 * temp2[3]);
+    ret(2, 2) = complex<real>(0.0, -2.0 * rsqrt_3 * temp2[3]);
     ret(0, 1) = complex<real>(temp1[0], temp1[1]);
     ret(1, 0) = complex<real>(-temp1[0], temp1[1]);
     ret(0, 2) = complex<real>(temp1[3], temp2[0]);
